refactor: constants 통합
