--- src/js/model/constants.js
@@ -1,1 +1,0 @@
-export const LOTTO_DIGIT = 6;
